requirements.txt
gunicorn
flask
google-generativeai
supabase
markdown
rapidfuzz
python-dotenv
werkzeug
